requirements.txt
streamlit==1.35.0
pandas==2.2.2
plotly==5.22.0
bcrypt==4.1.3
supabase==2.4.6
python-dotenv==1.0.1
